pyproject.toml
[project]
name = "thsr-ticket"
version = "0.2.0"
description = "An automatic booking program for Taiwan High Speed Railway (THSR)."
requires-python = ">=3.10"
dependencies = [
    "requests>=2.21",
    "pillow>=7.0",
    "beautifulsoup4>=4.8.2",
    "tinydb>=3.15.2",
    "pydantic<2.0",
    "opencv-python-headless>=4.0",
    "onnxruntime>=1.10",
    "numpy>=1.20",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "mypy>=0.76",
    "flake8>=3.7.9",
    "pylint>=2.4.4",
    "pytest>=5.3.5",
]
train = [
    "torch>=2.0",
]

[project.scripts]
thsr-ticket = "thsr_ticket.main:main"

[tool.setuptools.package-data]
thsr_ticket = ["ml/models/*.onnx", "ml/models/*.data"]
